% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy.R
\name{Entropy}
\alias{Entropy}
\title{Create and Plot Planck, Wein and Kircjhoff data}
\usage{
Entropy(Data, Field, GridSpace, GridType = "Hex", mode = 1)
}
\arguments{
\item{Data}{sf object}

\item{Field}{variable used to calculate entropy}

\item{GridSpace}{size in meters of the segmentation grid}

\item{GridType}{geometry of the grid (Sqr or Hex)}

\item{mode}{execute a final intersection with the provided Field}
}
\value{
Sf object
}
\description{
This functions creates the constant values
}
\keyword{Entropy}
